+++ bootcfg/storage/testfakes/fixtures.go
@@ -7,18 +7,18 @@ import (
 var (
 	// Group is a machine group for testing.
 	Group = &storagepb.Group{
-		Id:           "test-group",
-		Name:         "test group",
-		Profile:      "g1h2i3j4",
-		Requirements: map[string]string{"uuid": "a1b2c3d4"},
-		Metadata:     []byte(`{"k8s_version":"v1.1.2","pod_network":"10.2.0.0/16","service_name":"etcd2"}`),
+		Id:       "test-group",
+		Name:     "test group",
+		Profile:  "g1h2i3j4",
+		Selector: map[string]string{"uuid": "a1b2c3d4"},
+		Metadata: []byte(`{"k8s_version":"v1.1.2","pod_network":"10.2.0.0/16","service_name":"etcd2"}`),
 	}
 
 	// GroupNoMetadata is a Group without any metadata.
 	GroupNoMetadata = &storagepb.Group{
-		Id:           "group-no-metadata",
-		Requirements: map[string]string{"uuid": "a1b2c3d4"},
-		Metadata:     nil,
+		Id:       "group-no-metadata",
+		Selector: map[string]string{"uuid": "a1b2c3d4"},
+		Metadata: nil,
 	}
 
 	// Profile is a machine profile for testing.